--- a/seed/data_importer/match.py
+++ b/seed/data_importer/match.py
@@ -742,15 +742,15 @@ def link_states(states, ViewClass, cycle, highest_ali, sub_progress_key):  # noq
     else:
         state_class_name = "TaxLotState"
 
-    linked_view_ids = []
-    unlinked_view_ids = []
+    linked_views = []
+    unlinked_views = []
     invalid_link_states = []
     unlinked_states = []
 
     batch_size = math.ceil(len(states) / 100)
     for idx, state in enumerate(states):
         try:
-            _merge_count, link_count, view_id = match_merge_link(state.id, state_class_name, highest_ali=highest_ali, cycle=cycle)
+            _merge_count, link_count, view = match_merge_link(state, state_class_name, highest_ali=highest_ali, cycle=cycle)
         except (MultipleALIError, NoAccessError):
             invalid_link_states.append(state.id)
             continue
@@ -759,15 +759,12 @@ def link_states(states, ViewClass, cycle, highest_ali, sub_progress_key):  # noq
             continue
 
         if link_count == 0:
-            unlinked_view_ids.append(view_id)
+            unlinked_views.append(view)
         else:
-            linked_view_ids.append(view_id)
+            linked_views.append(view)
 
         if batch_size > 0 and idx % batch_size == 0:
             sub_progress_data.step("Matching Data (6/6): Merging Views")
-
-    linked_views = list(ViewClass.objects.filter(id__in=linked_view_ids))
-    unlinked_views = list(ViewClass.objects.filter(id__in=unlinked_view_ids))
 
     sub_progress_data.finish_with_success()